Mas ejercicios

+++ Capitulo03/C03Ejercicio04.java
@@ -30,8 +30,8 @@ public class C03Ejercicio04 {
     int multiplicacion;
     multiplicacion = primerNumero * segundoNumero;
     
-    int division;
-    division = primerNumero / segundoNumero;
+    double division;
+    division = ((double) primerNumero / (double)segundoNumero);
     
     System.out.print("El primer número introducido es " + primerNumero);
     System.out.println(" y el segundo es " + segundoNumero);
